分解Home组件中的create按钮
