+++ src/renderer/src/assets/css/index.css
@@ -125,14 +125,14 @@
     nav * {
         @apply z-30;
     }
-    nav .fab {
+    nav.fab {
         @apply fixed right-5 bottom-24 lg:bottom-5 z-40;
     }
-    nav .fab * {
+    nav.fab * {
         @apply z-40;
     }
     nav .navigation {
-        @apply fixed right-0 bottom-0 w-full lg:bottom-5 z-30;
+        @apply right-0 bottom-0 w-full lg:bottom-5 z-30;
     }
     nav .navigation div {
         @apply lg:rounded-3xl lg:mx-auto lg:max-w-lg lg:overflow-clip;
